PA-825 Fixed casting bug

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/v2/ui/dialog/DeclineDialog.java
@@ -210,7 +210,7 @@ public class DeclineDialog extends SimpleDialog {
     }
 
     private void onDeclined() {
-        _onDeclinedDispatcher.dispatch(getUid(), _workOrderId);
+        _onDeclinedDispatcher.dispatch(getUid(), (long) _workOrderId);
     }
 
     /*-*********************************-*/
